old Firefox web driver factory moved to specific package
