+++ core/worker/src/main/java/com/cognifide/aet/worker/drivers/firefox/FirefoxCommunicationWrapperImpl.java
